requirements.txt
streamlit  # ==1.32.0
deepl  # ==1.17.0
google-cloud-translate  # ==3.15.0
python-dotenv # ==1.0.1